requirements.txt
streamlit
ollama
sentence-transformers
numpy
torch
python-dotenv
